refactor files in persistence container, add services todo
