+++ services/Todo.md
@@ -1,0 +1,5 @@
+# manifesto model
+
+* per feature uncertainty correlations
+* per feature class correlations
+* limit prioritize endpoint samples to avoid sending back all
