+++ services/persistence/sqlite_wrapper.py
@@ -1,111 +1,19 @@
-# -*- coding: utf-8 -*-
-import json
-import os
 import sqlite3
-import requests
-
-from flask import Flask, jsonify, request
-
-app = Flask(__name__)
-
-DEBUG = True  # os.environ.get('DEBUG') is not None
-VERSION = 0.1
+import os
 
 DB_FILENAME = os.environ.get('DB_PATH')
 print('Using database', DB_FILENAME)
 
-MANIFESTO_MODEL_HTTP_PORT = os.environ.get('MANIFESTO_MODEL_HTTP_PORT')
-print('contact for manifesto model at port', MANIFESTO_MODEL_HTTP_PORT)
 
-
-@app.route("/", methods=['POST'])
-def index():
-    return jsonify({})
-
-
-@app.route("/training_texts", methods=['GET'])
-def training_texts():
-    """
-    Endpoint that returns the current state of the training data.
-    :return: {
-        'data': [
-            {],
-            ...
-        ]
-    }
-    """
-    texts = get_texts_with_labels(1000000)
-    texts = list(map(lambda entry: {'text': entry['statement'], 'label': entry['label']}, texts))
-    return jsonify({'data': texts}), 200
-
-
-@app.route("/texts_and_labels", methods=['POST'])
-def texts_and_labels():
-    """
-    Stores labels for the specified text ids.
-
-    expects a POST-body in the format:
-    {
-        "data": [
-            {"text_id": 17342, "label": "left"},
-            {"text_id": 873, "label": "neutral"},
-            ...
-        ]
-    }
-    """
-    texts_with_labels = json.loads(request.get_data(as_text=True))['data']
-    text_ids = map(lambda entry: entry['text_id'], texts_with_labels)
-    labels = map(lambda entry: entry['label'], texts_with_labels)
-
-    insert_into(DB_FILENAME, text_ids, labels, 'user')
-    n_inserts = len(texts_with_labels)
-
-    return jsonify({'n_inserted': n_inserts}), 201
-
-
-@app.route("/prioritized_texts", methods=['GET'])
-def prioritized_texts():
-    """
-    Retrieves all of the political statements from the database, sends them to the manifesto model to be prioritized
-    and returns at most as many texts given in the GET parameter `n`, ordered by their uncertainty.
-
-    :return: {
-        'data': [
-            {'text_id': 1, 'label': 'left'},
-            ...
-        ]
-    }
-    """
-    n_texts = int(request.args.get('n'))
-    texts = get_texts_only(1000000)
-
-    samples = {
-        'data': list(map(lambda entry: {'text_id': entry['text_id'], 'text': entry['statement']}, texts))
-    }
-
-    # ask model about sample uncertainty
-    url = 'http://manifesto_model:{}/estimate_uncertainty'.format(MANIFESTO_MODEL_HTTP_PORT)
-    r = requests.post(
-        url=url,
-        json=samples
-    )
-    priotized_text_ids = list(map(lambda e: int(e['text_id']), r.json()['data'][:n_texts]))
-    priotized_texts = get_texts_with_ids(priotized_text_ids)
-
-    text_data = {'data': priotized_texts}
-    return jsonify(text_data)
-
-
-def insert_into(database_filename, text_ids, labels, annotation_source):
+def insert_into(text_ids, labels, annotation_source):
     """
     inserts the texts and labels.
 
-    :param database_filename: name of the sqlite3 database.
     :param text_ids: iterable of int, ids of the texts.
     :param labels: iterable of int.
     :param annotation_source: string, manifesto or user.
     """
-    conn = sqlite3.connect(database_filename)
+    conn = sqlite3.connect(DB_FILENAME)
     c = conn.cursor()
 
     for text_id, label in zip(text_ids, labels):
@@ -233,13 +141,3 @@ def get_texts_with_labels(n_texts, label_strategy='duplicate'):
             )
         ]
         return results
-
-
-if __name__ == "__main__":
-    port = int(os.environ.get("HTTP_PORT"))
-    app.run(
-        host='0.0.0.0',
-        port=port,
-        debug=DEBUG,
-        use_reloader=False  # with reloader, caused main to be called twice
-    )
